phonebillapi/requirements.txt
Django==2.0.5
djangorestframework==3.8.2
psycopg2-binary
ipdb
python-dateutil
